test(agentic-ai): refactor e2e test setup into multiple test classes
