+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/AdHocToolsSchemaTests.java
@@ -16,6 +16,7 @@
  */
 package io.camunda.connector.e2e.agenticai;
 
+import static io.camunda.connector.e2e.agenticai.aiagent.AiAgentTestFixtures.AD_HOC_TOOLS_SCHEMA_ELEMENT_TEMPLATE_PATH;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.core.type.TypeReference;
