+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/langchain4j/Langchain4JAiAgentResponseHandlingTests.java
@@ -1,0 +1,181 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.camunda.connector.e2e.agenticai.aiagent.langchain4j;
+
+import static io.camunda.connector.e2e.agenticai.aiagent.AiAgentTestFixtures.AI_AGENT_TASK_ID;
+import static io.camunda.connector.e2e.agenticai.aiagent.AiAgentTestFixtures.HAIKU_JSON;
+import static io.camunda.connector.e2e.agenticai.aiagent.AiAgentTestFixtures.HAIKU_JSON_ASSERTIONS;
+import static io.camunda.connector.e2e.agenticai.aiagent.AiAgentTestFixtures.HAIKU_TEXT;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.camunda.connector.e2e.agenticai.assertj.AgentResponseAssert;
+import io.camunda.connector.test.SlowTest;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+
+@SlowTest
+public class Langchain4JAiAgentResponseHandlingTests extends BaseLangchain4JAiAgentTests {
+
+  @Nested
+  class ResponseText {
+
+    @Test
+    void fallsBackToResponseTextWhenNoResponsePropertiesAreConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate -> elementTemplate.withoutPropertyValueStartingWith("data.response."),
+          HAIKU_TEXT,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasNoResponseMessage()
+                  .hasResponseText(HAIKU_TEXT)
+                  .hasNoResponseJson());
+    }
+
+    @Test
+    void returnsResponseTextIfConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "text")
+                  .property("data.response.format.includeText", "=true")
+                  .property("data.response.includeAssistantMessage", "=false"),
+          HAIKU_TEXT,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasNoResponseMessage()
+                  .hasResponseText(HAIKU_TEXT)
+                  .hasNoResponseJson());
+    }
+
+    @Test
+    void returnsOnlyResponseMessageIfConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "text")
+                  .property("data.response.format.includeText", "=false")
+                  .property("data.response.includeAssistantMessage", "=true"),
+          HAIKU_TEXT,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasResponseMessageText(HAIKU_TEXT)
+                  .hasNoResponseText()
+                  .hasNoResponseJson());
+    }
+
+    @Test
+    void returnsBothResponseTextAndMessageIfConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "text")
+                  .property("data.response.format.includeText", "=true")
+                  .property("data.response.includeAssistantMessage", "=true"),
+          HAIKU_TEXT,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasResponseMessageText(HAIKU_TEXT)
+                  .hasResponseText(HAIKU_TEXT)
+                  .hasNoResponseJson());
+    }
+
+    @Test
+    void triesToParseTextResponseAsJsonIfConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "text")
+                  .property("data.response.format.includeText", "=true")
+                  .property("data.response.format.parseTextToJson", "=true"),
+          HAIKU_JSON,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasResponseText(HAIKU_JSON)
+                  .hasResponseJsonSatisfying(HAIKU_JSON_ASSERTIONS));
+    }
+
+    @Test
+    void returnsNullWhenTextResponseFailsToBeParsedAsJson() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "text")
+                  .property("data.response.format.includeText", "=true")
+                  .property("data.response.format.parseTextToJson", "=true"),
+          HAIKU_TEXT,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasResponseText(HAIKU_TEXT)
+                  .hasNoResponseJson());
+    }
+  }
+
+  @Nested
+  class ResponseJson {
+
+    @Test
+    void returnsJsonObject() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate -> elementTemplate.property("data.response.format.type", "json"),
+          HAIKU_JSON,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasNoResponseText()
+                  .hasResponseJsonSatisfying(HAIKU_JSON_ASSERTIONS));
+    }
+
+    @Test
+    void returnsBothResponseJsonAndMessageIfConfigured() throws Exception {
+      testBasicExecutionWithoutFeedbackLoop(
+          elementTemplate ->
+              elementTemplate
+                  .property("data.response.format.type", "json")
+                  .property("data.response.includeAssistantMessage", "=true"),
+          HAIKU_JSON,
+          true,
+          (agentResponse) ->
+              AgentResponseAssert.assertThat(agentResponse)
+                  .hasResponseMessageText(HAIKU_JSON)
+                  .hasNoResponseText()
+                  .hasResponseJsonSatisfying(HAIKU_JSON_ASSERTIONS));
+    }
+
+    @Test
+    void raisesIncidentWhenJsonCouldNotBeParsed() throws Exception {
+      final var setup =
+          setupBasicTestWithoutFeedbackLoop(
+              elementTemplate -> elementTemplate.property("data.response.format.type", "json"),
+              HAIKU_TEXT);
+      setup.getRight().waitForActiveIncidents();
+
+      assertIncident(
+          setup.getRight(),
+          incident -> {
+            assertThat(incident.getElementId()).isEqualTo(AI_AGENT_TASK_ID);
+            assertThat(incident.getErrorMessage())
+                .startsWith("Failed to parse response content as JSON");
+          });
+    }
+  }
+}
